requirements.txt
python-dotenv~=1.0.1
robotframework~=7.0
robotframework-seleniumlibrary~=6.3.0
webdriver-manager